Cput

+++ src/mcw_malloc.c
@@ -91,7 +91,7 @@ static UINT       serial = 0    ; /* serial number of allocation */
 
 static INLINE UINT mallkey( char * fred )
 {
-   UINT q = ((UINT) fred) ;
+   UINT q = (UINT)(intptr_t)fred ;
 
    q =   ((q & 0xf0f0f0f0) >> 4)   /* swap nibbles */
        | ((q & 0x0f0f0f0f) << 4) ;
